move wcRequest to walletConnectStore
