+++ src/stores/walletconnectStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia"
 import { ref } from 'vue'
 import { Core } from '@walletconnect/core'
-import { Web3Wallet } from '@walletconnect/web3wallet'
+import { Web3Wallet, type Web3WalletTypes } from '@walletconnect/web3wallet'
 import type Client from '@walletconnect/web3wallet'
 import type {SessionTypes} from '@walletconnect/types'
 
@@ -29,6 +29,48 @@ export const useWalletconnectStore = defineStore('walletconnectStore', () => {
     activeSessions.value = web3wallet.value.getActiveSessions();
   }
   
+  // Wallet connect dialog functionality
+  async function wcRequest(event: Web3WalletTypes.SessionRequest, walletAddress: string) {
+    if(!web3wallet.value) throw new Error("No web3wallet initialized")
+    const { topic, params, id } = event;
+    const { request } = params;
+    const method = request.method;
 
-  return { web3wallet, activeSessions, initweb3wallet }
+    switch (method) {
+      case "bch_getAddresses":
+      case "bch_getAccounts": {
+        const result = [walletAddress];
+        const response = { id, jsonrpc: '2.0', result };
+        web3wallet.value.respondSessionRequest({ topic, response });
+      }
+        break;
+      case "bch_signMessage":
+      case "personal_sign": {
+        const sessions = web3wallet.value.getActiveSessions();
+        const session = sessions[topic];
+        if (!session) return;
+        const metadataDapp = session.peer.metadata;
+        // TODO: rework
+        //dappMetadata.value = metadataDapp
+        //signMessageRequestWC.value = event;
+      }
+        break;
+      case "bch_signTransaction": {
+        const sessions = web3wallet.value.getActiveSessions();
+        const session = sessions[topic];
+        if (!session) return;
+        const metadataDapp = session.peer.metadata;
+        // TODO: rework
+        //dappMetadata.value = metadataDapp
+        //transactionRequestWC.value = event;
+      }
+        break;
+      default:{
+        const response = { id, jsonrpc: '2.0', error: {code: 1001, message: `Unsupported method ${method}`} };
+        await web3wallet.value.respondSessionRequest({ topic, response });
+      }
+    }
+  }
+
+  return { web3wallet, activeSessions, initweb3wallet, wcRequest }
 })
